Modified ScafTranscripts.pl to add info about skipped contigs to log and to avoid unnecessary seq lookups
Contigs that have 45 bp or more of non homologous overlap with their top hit (ie; alignment
starts/ends >45 bp from start/end of both query and hit) are not scaffolded but are simply
written to the outfile. Previously, this was not noted in the log. They were indistinguishable
from singleton contigs. I modified the log to indicate that these contigs have too much
non-homologous sequence.

I also moved the lookup of the contig sequence to right before it is written to file
because many contigs were being retrieved but not written. This was necessary Previously
because I had to look up the length of the sequence to determine the overlap. This information
is now provided in the blast result, so the lookups were unnecessary at that stage.