Modified to allow multiple consensus sequences from multiple monophyletic groups
This involved replacing several functions that I wrote to test for monophyly of all
sequences from a species with the ete2 get_monophyletic method which was not available
in the version that I've been using up until now.

I also had to annotate the trees by adding species info as a feature. That way all leaves
from the same species have a common annotation that can be fed to get_monophyletic

I also modified the script to work on sequences from a single family, which is passed in
from the command line using the -s option.

I also fixed the potential problem of multiple consensus sequences with the same name
being produced (at least within a single cluster, there could still be sequences from
different clusters with the same name.