Removed the functionality of removing sequences from cluster fasta files and replacing them with consensus sequences
I was not happy about having this (somewhat unrelated) functionality in this script.
The only reason it was placed here was because I need the info about which cluster each
consensus sequence belongs to. This info is now stored in the GTF file, which I am also
going to use to extract the coding portion of the sequence and add it to the corresponding
cluster as a separate step