convert numeric values from blast output to int/float and test both subject and query for reversed coordinates
annoyingly, blastx and blastn differ in which sequence is reversed when the hit is on the
opposite strand. This script now tests for both. As long as both aren't reversed (indicating
the same strand), this should work

This issue turned out to be somewhat moot, because the script was doing a lexical comparison
of values. This is the second time this "feature" of python has bitten me in the ass. I
have now converted all numeric fields from the blast to int or float