Corrected problem with smart_consensus where Ns were being added instead of gaps in cases all sequences had Ns at a position
this was caused by situations where some sequences aligned entirely before an alignment
position and all others aligned completely after it.

This caused Ns to be inserted instead of gaps. Since there were non-Ns before and after,
these were not converted to gaps when pad ends was run on the consensus sequence.
I've now fixed this so that gaps are added instead of Ns (so it's no longer necessary to
run pad ends on the consensus sequence). However, this does mean that if there are truly
Ns in all of the sequences at an internal alignment position, that position will be turned
into a gap. This is probably an unlikely situation and wouldn't make all that big a
difference anyway (though it might screw up the coding frame)