Added function to identify C-terminal CaaX domains in DNA sequences
This motif is characteristic of the pheromone genes in basidiomycete B locus mating type
loci. It can be used to identify pheromone genes on the same contig as the pheromone
receptor gene

In addition to find_CaaX, I also wrote a function called six_frame_translation that is
used by find_CaaX to translate a DNA sequence in all possible reading frames.

find_CaaX could be fairly easily generalized to find any sequence motif passed to the
function, but I was in a hurry so I didn't bother to write it this way