Refactored to simplify flow control, added taxon deduplication
Previous versions had different input code depending on the output options.
This made for convoluted logic and made changing anything difficult.
This structure was motivated by wanting to use the convert option in alignIO as much as
possible but that is a minor benefit that came at considerable cost.
The code now simply reads in the alignment depending on the input options, then outputs
it based on the output options. Simple.
This also makes it easy to manipulate alignments between input and output, such as
removing sequences with identical names.